Mods to PEUC section to correspond to literature.
The Pacific EUC section now goes from -2 to 2 latitude, in order
to agree with the Sloyan et al citation used for the observational
estimate.

The Agulhas section has been extended to allow for its use
across more models.  This change is not going to modify diagnostic
answers, since the extension is into land.