SIS2: *Provide ocean SST to atmosphere on first step
  For some reason (perhaps it is a bug), the ice model had been replacing the
SST from the ocean_surface_state with an SST read in from a data override via
get_sea_surface during the first step of a run that was not initialized from a
restart file.  This seems like a bad idea (M. Winton concurs), so we are
changing this so that the SST from the ocean is always provided to the
atmosphere, even on the first step.  This single line correction changes answers
at the very beginning of any run without a restart file, but the solutions are
unchanged thereafter or if a restart file is used.  New reference solutions have
been checked in (except for the coupled models which are started with an
ice_model.res.nc.  Given the nature of this problem we decided not to provide a
flag that recovers the previous answers.
SIS2 commit NOAA-GFDL/SIS2@1b068886d000c4bfbcfd3b900f5285629adf0458