SIS2: *Corrected a bug with FILLING_FRAZIL=True
  Corrected a signficant bug in the application of the frazil heat flux when
FILLING_FRAZIL=True.  The specific problem was a missing backward increment in a
loop, but it was leading to a significant non-conservation of heat and the
creation of far too little sea ice.  Also use the minimum of the actual ice
thickness and the category lower bound when calculating the relative enthalpy
extraction by frazil; this minor change mostly affects categories that are set
to have both 0 thickness and 0 fractional area if there is another thicker
category with nonzero area. Because FILLING_FRAZIL=True by default, this leads
to answer changes, and new reference solutions have been checked in.
NOAA-GFDL/SIS2@40725813242d9d593de34b5ddef9d88eb0d286a4