*Eliminate CORRECT_DENSITY from MOM_input files
  Removed lines setting CORRECT_DENSITY in all MOM_input files, which is the
same as setting CORRECT_DENSITY=True in cases that use MOM_entrain_diffusive,
and is eliminating an unused parameter in other cases.  This changes answers
in only one experiment, resting/layer, which uses CORRECT_DENSITY=False with
a non-zero value of Kd.  The MOM_parameter_doc files are changed in 7 other
examples, but the answers do not change.

- NOAA-GFDL/MOM6-examples: 5c9f52b *Eliminate CORRECT_DENSITY from MOM_input files