C5 updates
The environment files have been updated to run on the C5 cluster.  One
notable change is that "PGI" has been replaced with Nvidia settings, and
the PGI compiler is spoofed.

The MOM6-examples submodules has been updated to point to a
C5-compatible repository.

.gitignore has been updated to include the scratch (tmp-*) directories
used to generate output.

NOTE: This commit does not update the answers for C5.  These are to
appear in a separate commit.